Quebrando os exemplos em 3 arquivos
